feat(dashboard): extract session list component

Aloop-Agent: build
Aloop-Iteration: 14
Aloop-Session: orchestrator-20260321-172932-issue-154-20260321-181044
