+++ aloop/cli/dashboard/src/AppView.tsx
@@ -9,8 +9,8 @@ import {
 } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/components/ui/collapsible';
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
 import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from '@/components/ui/command';
 import { HoverCard, HoverCardContent, HoverCardTrigger } from '@/components/ui/hover-card';
 import { Progress } from '@/components/ui/progress';
@@ -22,6 +22,7 @@ import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/comp
 import { parseTodoProgress } from '../../src/lib/parseTodoProgress';
 import { relativeTime, phaseColors, phaseBarColors, phaseDotColors, statusColors, PhaseBadge, STATUS_DOT_CONFIG, StatusDot } from '@/components/session/helpers';
 import { SessionCard } from '@/components/session/SessionCard';
+import { SessionList, type SessionListSession } from '@/components/session/SessionList';
 export { relativeTime } from '@/components/session/helpers';
 
 // ── ANSI + Markdown rendering ──
@@ -187,6 +188,10 @@ interface SessionSummary {
   workDir: string;
   stuckCount: number;
 }
+
+const asSessionListSession = (session: SessionSummary): SessionListSession => ({
+  ...session,
+});
 
 interface LogEntry {
   timestamp: string;
@@ -580,33 +585,7 @@ export function Sidebar({
   collapsed: boolean;
   onToggle: () => void;
 }) {
-  // Group by project
-  const { projectGroups, olderSessions } = useMemo(() => {
-    const now = Date.now();
-    const cutoff = 24 * 60 * 60 * 1000; // 24h
-    const active: SessionSummary[] = [];
-    const older: SessionSummary[] = [];
-
-    for (const s of sessions) {
-      const lastActivity = s.endedAt || s.startedAt;
-      const age = lastActivity ? now - new Date(lastActivity).getTime() : Infinity;
-      if (s.isActive || s.status === 'running' || age < cutoff) {
-        active.push(s);
-      } else {
-        older.push(s);
-      }
-    }
-
-    const groups = new Map<string, SessionSummary[]>();
-    for (const s of active) {
-      const key = s.projectName || 'Unknown';
-      if (!groups.has(key)) groups.set(key, []);
-      groups.get(key)!.push(s);
-    }
-    return { projectGroups: groups, olderSessions: older };
-  }, [sessions]);
-
-  const [olderOpen, setOlderOpen] = useState(false);
+  const listSessions = useMemo(() => sessions.map(asSessionListSession), [sessions]);
 
   useEffect(() => {
     if (!contextMenuSessionId) return;
@@ -649,9 +628,6 @@ export function Sidebar({
     );
   }
 
-  const isSelected = (s: SessionSummary) =>
-    selectedSessionId === null ? sessions.indexOf(s) === 0 : s.id === selectedSessionId;
-
   return (
     <aside className="flex flex-col border-r border-border bg-sidebar w-64 shrink-0 animate-slide-in-left">
       <div className="flex items-center justify-between px-3 py-2 border-b border-border">
@@ -665,98 +641,7 @@ export function Sidebar({
           <TooltipContent><p>Collapse (Ctrl+B)</p></TooltipContent>
         </Tooltip>
       </div>
-      <ScrollArea className="flex-1">
-        <div className="p-2 overflow-hidden">
-          {Array.from(projectGroups.entries()).map(([project, items]) => (
-            <Collapsible key={project} defaultOpen>
-              <CollapsibleTrigger className="flex items-center gap-1 w-full px-1 py-1 text-[10px] font-semibold text-muted-foreground/70 uppercase tracking-wider hover:text-muted-foreground">
-                <ChevronDown className="h-3 w-3 transition-transform group-data-[state=closed]:rotate-[-90deg]" />
-                {project}
-                <span className="ml-auto text-muted-foreground/40">{items.length}</span>
-              </CollapsibleTrigger>
-              <CollapsibleContent>
-                <div className="space-y-0.5 mb-2">
-                  {items.map((session) => (
-                    <SessionCard
-                      key={session.id}
-                      session={session}
-                      selected={isSelected(session)}
-                      onSelectSession={onSelectSession}
-                    />
-                  ))}
-                </div>
-              </CollapsibleContent>
-            </Collapsible>
-          ))}
-
-          {olderSessions.length > 0 && (
-            <Collapsible open={olderOpen} onOpenChange={setOlderOpen}>
-              <CollapsibleTrigger className="flex items-center gap-1 w-full px-1 py-1 text-[10px] font-semibold text-muted-foreground/40 uppercase tracking-wider hover:text-muted-foreground">
-                {olderOpen ? <ChevronDown className="h-3 w-3" /> : <ChevronRight className="h-3 w-3" />}
-                Older
-                <span className="ml-auto">{olderSessions.length}</span>
-              </CollapsibleTrigger>
-              <CollapsibleContent>
-                <div className="space-y-0.5 mb-2">
-                  {olderSessions.map((session) => (
-                    <SessionCard
-                      key={session.id}
-                      session={session}
-                      selected={isSelected(session)}
-                      onSelectSession={onSelectSession}
-                    />
-                  ))}
-                </div>
-              </CollapsibleContent>
-            </Collapsible>
-          )}
-
-          {sessions.length === 0 && <p className="text-xs text-muted-foreground p-2">No sessions.</p>}
-        </div>
-      </ScrollArea>
-      {contextMenuSessionId && (
-        <div
-          role="menu"
-          className="fixed z-50 min-w-[170px] rounded-md border bg-popover p-1 text-popover-foreground shadow-md"
-          style={{ left: contextMenuPos.x, top: contextMenuPos.y }}
-          onPointerDown={(event) => event.stopPropagation()}
-        >
-          <button
-            type="button"
-            className="w-full flex items-center gap-2 rounded-sm px-2 py-1.5 min-h-[44px] md:min-h-0 text-sm text-left hover:bg-accent"
-            onClick={() => {
-              const selectId = contextMenuSessionId === 'current' ? null : contextMenuSessionId;
-              onSelectSession(selectId);
-              onStopSession?.(selectId, false);
-              setContextMenuSessionId(null);
-            }}
-          >
-            <Square className="h-3.5 w-3.5" /> Stop after iteration
-          </button>
-          <button
-            type="button"
-            className="w-full flex items-center gap-2 rounded-sm px-2 py-1.5 min-h-[44px] md:min-h-0 text-sm text-left text-destructive hover:bg-accent"
-            onClick={() => {
-              const selectId = contextMenuSessionId === 'current' ? null : contextMenuSessionId;
-              onSelectSession(selectId);
-              onStopSession?.(selectId, true);
-              setContextMenuSessionId(null);
-            }}
-          >
-            <Zap className="h-3.5 w-3.5" /> Kill immediately
-          </button>
-          <button
-            type="button"
-            className="w-full flex items-center gap-2 rounded-sm px-2 py-1.5 min-h-[44px] md:min-h-0 text-sm text-left hover:bg-accent"
-            onClick={() => {
-              onCopySessionId?.(contextMenuSessionId);
-              setContextMenuSessionId(null);
-            }}
-          >
-            <GitCommit className="h-3.5 w-3.5" /> Copy session ID
-          </button>
-        </div>
-      )}
+      <SessionList sessions={listSessions} selectedSessionId={selectedSessionId} onSelectSession={onSelectSession} />
     </aside>
   );
 }
